separate bookDetails component to one directory
